+++ docs/md/Version-history.md
@@ -24,7 +24,8 @@ Available in [pre-release](https://www.sumatrapdfreader.org/prerelease) builds.
 - Can save images in different formats: PNG, JPEG, BMP, GIF, TIFF.
 - add `Show Errors` in right-click context menu for PDF documents that have mupdf warnings/errors
 - add `CmdToggleSmoothScroll` (`Toggle Smooth Scroll`) command to toggle `SmoothScroll` advanced setting
-- add `CmdToggleHideScrollbar` (`Toggle Hide Scrollbar`) command to toggle `HideScrollbar` advanced setting
+- replace `HideScrollbars` and `UseOverlayScrollbar` settings with `Scrollbars` setting (values: `windows`, `smart`, `overlay`, `hidden`)
+- add `CmdChangeScrollbar` (`Change Scrollbar`) command to open scrollbar mode dialog
 - add `CmdToggleScrollbarInSinglePage` (`Toggle Scrollbar In Single Page`) command to toggle `ScrollbarInSinglePage` advanced setting
 - add `TabsMru` advanced setting and `CmdToggleTabsMru` (`Toggle Tabs MRU`) command to toggle it. It changes order of navigating tabs when usint `Ctrl + Tab` (`CmdNextTabSmart`)
 - improve document properties for comic book files (CBZ, CBR, CB7, CBT). We now show list of image files.